+++ .github/workflows/main/installmtoa.sh
@@ -61,8 +61,12 @@ fi
 
 mkdir -p mtoaRoot/$mtoaVersion && cd mtoaRoot/$mtoaVersion
 
-echo Downloading MtoA "https://${url}"
-curl -L https://${login}${url} -o mtoa-${mtoaVersion}-${arnoldPlatform}-${mayaVersion}.run
+if [ !  -e pipeline/build/.download/mtoa-${mtoaVersion}-${arnoldPlatform}-${mayaVersion}.run ] ; then
+	echo Downloading MtoA "https://${url}"
+	curl -L https://${login}${url} -o mtoa-${mtoaVersion}-${arnoldPlatform}-${mayaVersion}.run
+	mv mtoa-${mtoaVersion}-${arnoldPlatform}-${mayaVersion}.run ../../pipeline/build/.download/ || exit -1
+fi
+ln -s ../../pipeline/build/.download/mtoa-${mtoaVersion}-${arnoldPlatform}-${mayaVersion}.run ./
 
 if [ "$DOCKER_PYTHON" == "" ] && [ "$TRAVIS" == "" ] ; then
 	sh ./mtoa-${mtoaVersion}-${arnoldPlatform}-${mayaVersion}.run --tar xvf
